+++ src/components/header/MenuBlock.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function MenuBlock() {
+  return <div>MenuBlock</div>;
+}
+
+export default MenuBlock;
